backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
django-storages
boto3
